Define relationship between device & user
Device model is expanded with the explicitly defined foreign key to the
user's table, and the access key field which is not yet initialized.
Note that the reference integrity check relies on checking the error
message instead of the error code due to the bug present in the
CockroachDB code.

Signed-off-by: Dejan Mijic <dejan@mainflux.com>